requirements.txt
agno==1.2.11
fastapi==0.110.1
uvicorn==0.29.0
pydantic-settings==2.8.1
groq==0.22.0
qdrant-client==1.13.3
fastembed==0.6.0
requests==2.32.3
python-dotenv==1.1.0
python-multipart==0.0.20